Refactor PrimaryMenu structure
