+++ TrinityFrontend/src/components/PrimaryMenu/AppIdentity/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AppIdentity';
