+++ TrinityFrontend/src/components/PrimaryMenu/MyProfile/ProfileMenu/components/UserInfo.tsx
