+++ TrinityFrontend/src/components/PrimaryMenu/MyProfile/SettingsMenu/components/Integrations.tsx
