+++ TrinityFrontend/src/components/PrimaryMenu/MyProfile/SettingsMenu/components/WorkspacePreferences.tsx
