+++ TrinityFrontend/src/components/PrimaryMenu/MyProfile/SettingsMenu/index.ts
