+++ TrinityFrontend/src/components/PrimaryMenu/MyProfile/index.ts
@@ -1,0 +1,3 @@
+export { default as NotificationMenu } from './NotificationMenu';
+export { default as SettingsMenu } from './SettingsMenu';
+export { default as ProfileMenu } from './ProfileMenu';
