+++ TrinityFrontend/src/components/PrimaryMenu/Searchbar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Searchbar';
