+++ TrinityFrontend/src/components/PrimaryMenu/TrinityAssets/AnimatedLogo.tsx
