+++ TrinityFrontend/src/components/PrimaryMenu/TrinityAssets/BackToAppsIcon.tsx
