+++ TrinityFrontend/src/components/PrimaryMenu/TrinityAssets/index.ts
@@ -1,0 +1,2 @@
+export { default as AnimatedLogo } from './AnimatedLogo';
+export { default as BackToAppsIcon } from './BackToAppsIcon';
