+++ TrinityFrontend/src/components/PrimaryMenu/index.ts
@@ -1,1 +1,5 @@
-export { default } from './PrimaryMenu';
+export * as TrinityAssets from './TrinityAssets';
+export { default as AppIdentity } from './AppIdentity';
+export { default as Searchbar } from './Searchbar';
+export * as MyProfile from './MyProfile';
+export { default as Navigation } from './Navigation';
